// be/src/config/configKeys.enum.ts
export enum ConfigKeys {
  NODE_ENV = 'NODE_ENV',
  PORT = 'PORT',
  FE_PORT = 'FE_PORT',
  BE_BASE_URL = 'BE_BASE_URL',
  FE_BASE_URL = 'FE_BASE_URL',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_USER = 'DB_USER',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_NAME = 'DB_NAME',
  CLIENT_ID_LINKEDIN = 'CLIENT_ID_LINKEDIN',
  CLIENT_SECRET_LINKEDIN = 'CLIENT_SECRET_LINKEDIN',
  JWT_SECRET = 'JWT_SECRET',
}